+++ scripts/winesapos-install.sh
@@ -850,6 +850,9 @@ echo "Setting up the desktop environment complete."
 echo 'Setting up the additional package managers...'
 yay_install_chroot appimagepool-appimage bauh snapd
 chroot ${WINESAPOS_INSTALL_DIR} systemctl enable snapd
+# Enable support for classic Snaps.
+mkdir -p ${WINESAPOS_INSTALL_DIR}/var/lib/snapd/snap
+ln -s /var/lib/snapd/snap ${WINESAPOS_INSTALL_DIR}/snap
 
 if [[ "${WINESAPOS_DISTRO_DETECTED}" == "manjaro" ]]; then
     chroot ${WINESAPOS_INSTALL_DIR} ${CMD_PACMAN_INSTALL[*]} appimagelauncher